+++ .eslintrc.js
@@ -1,6 +1,7 @@
 module.exports = {
   'env': {
     'browser': true,
+    'node': true,
     'es2021': true,
   },
   'extends': 'google',
