+++ public/components/card/card.css
@@ -1,0 +1,27 @@
+.card-body {
+
+}
+
+.card-img {
+
+}
+
+.card-info {
+
+}
+
+.card-fav-btn {
+
+}
+
+.card-fav-btn:hover {
+        
+}
+
+/* .card-fav-btn:active {
+
+} */
+
+.badge {
+        
+}
