+++ public/components/header/header.hbs
@@ -1,0 +1,1 @@
+{{!-- structure --}}
