+++ public/components/icons/search.mjs
@@ -1,0 +1,8 @@
+export default function search() {
+        const root = document.createElement('div')
+        root.innerHTML = `<svg xmlns="http://www.w3.org/2000/svg" width="17" height="16" viewBox="0 0 17 16" fill="none">
+        <path d="M7.00006 1C10.0377 1 12.5001 3.46256 12.5001 6.50029C12.5001 7.74868 12.0842 8.89993 11.3835 9.82304L15.2791 13.7233C15.5718 14.0164 15.5715 14.4913 15.2785 14.784C14.9854 15.0767 14.5105 15.0764 14.2178 14.7834L10.3227 10.8839C9.39959 11.5847 8.24839 12.0006 7.00006 12.0006C3.96246 12.0006 1.5 9.53802 1.5 6.50029C1.5 3.46256 3.96246 1 7.00006 1ZM7.00006 2.5C4.7909 2.5 3 4.29098 3 6.50029C3 8.70961 4.7909 10.5006 7.00006 10.5006C9.20921 10.5006 11.0001 8.70961 11.0001 6.50029C11.0001 4.29098 9.20921 2.5 7.00006 2.5Z" fill="#919399"/>
+        </svg>`
+        root.classList = ['svgIcon']
+        return root
+}
