+++ public/index.html
@@ -5,7 +5,7 @@
     <link rel="stylesheet" href="/css/styles.css">
 </head>
 <body>
-    <!-- <img style="width: auto; height: 100vh;" src="https://a-static.besthdwallpaper.com/cyberpunk-2077-video-game-cyborg-bus-stop-wallpaper-2880x1800-63164_8.jpg" alt=""> -->
+    <script src="https://cdn.jsdelivr.net/npm/handlebars@latest/dist/handlebars.js"></script>
     <div id="root"></div>
     <script type="module" src="./main.js"></script>
 </body>
